refactor: move the code that generates content to a new file
